xhrupload: Improve `getResponseError` handling
When returning a string from `getResponseError`, it will be wrapped in
an Error instance.

When returning a falsy value from `getResponseError`, a default Error
message is used.

When returning something completely different, an Error instance is
created and its `.data` property will equal the return value.

Errors now include a `.request` property, whether they were caused by
the request failing entirely ('error' event) or the backend rejecting
it (during the 'load' event). Previously this only happened in the
second case.

Ref https://github.com/transloadit/uppy/issues/521#issuecomment-357612472